Merge pull request #255 from macarooni-man/dev

+++ source/core/server/backup.py
@@ -9,6 +9,9 @@ from source.core.translator import translate
 from source.core.constants import paths
 from source.core.server import manager
 from source.core import constants
+from source.core.constants import (
+    load_config
+)
 
 
 # Auto-MCS Back-up API
@@ -44,19 +47,15 @@ class BackupObject():
         cfg_list.extend(glob(os.path.join(extract_folder, '.auto-mcs.ini')))
 
         for cfg in cfg_list:
-            config = ConfigParser(allow_no_value=True, comment_prefixes=';', interpolation=None)
-            config.optionxform = str
-            config.read(cfg)
-            if config:
+            config = load_config(cfg)
+            if config.sections():
 
                 # If auto-mcs.ini exists, grab version and type information
                 if config.get('general', 'serverName') == self.name:
                     self.type = config.get('general', 'serverType')
                     self.version = config.get('general', 'serverVersion')
-                    try:
-                        self.build = config.get('general', 'serverBuild')
-                    except:
-                        self.build = None
+                    try:    self.build = config.get('general', 'serverBuild')
+                    except: self.build = None
 
                 os.remove(cfg)
                 break
@@ -145,13 +144,13 @@ class BackupManager():
     # Backup functions
 
     # Backs up server to the backup directory in auto-mcs.ini
-    def save(self, ignore_running=False):
+    def save(self, ignore_running=False) -> dict[str, str] | None:
         backup = backup_server(self._server['name'], self._backup_stats, ignore_running)
         self._update_data()
         return backup
 
     # Restores server from file name
-    def restore(self, backup_obj: BackupObject):
+    def restore(self, backup_obj: BackupObject) -> dict[str, str] | None:
         if self._server['name'] not in constants.server_manager.running_servers:
             backup = restore_server(self._server['name'], backup_obj.path, self._backup_stats)
             self._update_data()
@@ -295,9 +294,9 @@ def dump_config(server_name: str, new_server=False):
         # Only pickup server as valid with good config
         if server_name == server_config.get("general", "serverName"):
             server_dict['version'] = server_config.get("general", "serverVersion")
-            backup_stats['backup-path'] = str(server_config.get("bkup", "bkupDir"))
-            backup_stats['auto-backup'] = str(server_config.get("bkup", "bkupAuto").lower())
-            backup_stats['max-backup'] = str(server_config.get("bkup", "bkupMax"))
+            backup_stats['backup-path'] = str(server_config.get("bkup", "bkupDir",  fallback = paths.backups))
+            backup_stats['auto-backup'] = str(server_config.get("bkup", "bkupAuto", fallback = 'false')).lower()
+            backup_stats['max-backup']  = str(server_config.get("bkup", "bkupMax",  fallback = '5'))
 
 
     # Generate backup list and metadata
@@ -322,7 +321,7 @@ def dump_config(server_name: str, new_server=False):
 # ---------------------------------------------- Backup Functions ------------------------------------------------------
 
 # name --> backup to directory
-def backup_server(name: str, backup_stats=None, ignore_running=False):
+def backup_server(name: str, backup_stats=None, ignore_running=False) -> dict[str, str] | None:
 
     if set_lock(name, True, 'save'):
 
@@ -352,8 +351,11 @@ def backup_server(name: str, backup_stats=None, ignore_running=False):
             constants.folder_check(backup_path)
             server_path = manager.server_path(name)
 
+
+            # Attempt to save the backup
             failed = True
-            while failed:
+            total_attempts = 3
+            for attempt in range(total_attempts):
                 if os.path.exists(temp_backup):
                     constants.safe_delete(temp_backup)
                     constants.folder_check(temp_backup)
@@ -366,7 +368,17 @@ def backup_server(name: str, backup_stats=None, ignore_running=False):
                 if code != 0: continue
 
                 failed = False
+                break
 
+
+            # Log and exit with failure
+            if failed:
+                set_lock(name, False)
+                send_log('backup_server', f"failed to save a back-up of '{name}' after {total_attempts} attempts", 'error')
+                return
+
+
+            # Success, continue
             if os.path.exists(backup_file): os.remove(backup_file)
             os.rename(os.path.join(backup_path, f"{name}-bkup.amb"), backup_file)
 
@@ -392,23 +404,24 @@ def backup_server(name: str, backup_stats=None, ignore_running=False):
 
             set_lock(name, False)
             send_log('backup_server', f"successfully saved a back-up of '{name}' to '{backup_file}'", 'info')
-            return [file_name, convert_size(os.stat(backup_file).st_size), bkup_time]
+            return {
+                'name': file_name,
+                'size': convert_size(os.stat(backup_file).st_size),
+                'date': bkup_time
+            }
 
         except Exception as e:
             send_log('backup_server', f"error saving a back-up of '{name}': {constants.format_traceback(e)}", 'error')
 
 
 # name, index --> restore from file
-def restore_server(name: str, backup_name: str, backup_stats=None):
+def restore_server(name: str, backup_name: str, backup_stats=None) -> dict[str, str] | None:
 
     if set_lock(name, True, 'restore'):
 
         try:
             send_log('restore_server', f"restoring '{name}' to '{backup_name}'...", 'info')
-            constants.java_check()
-
-            if not backup_stats:
-                backup_stats = dump_config(name)[1]
+            if not backup_stats: backup_stats = dump_config(name)[1]
 
             cwd = constants.get_cwd()
             backup_path = backup_stats["backup-path"]
@@ -428,15 +441,12 @@ def restore_server(name: str, backup_name: str, backup_stats=None):
                 # Delete all files in server directory
                 for f in glob(os.path.join(manager.server_path(name), '*')):
                     try:
-                        if os.path.isdir(f):
-                            constants.safe_delete(f)
-                        else:
-                            os.remove(f)
+                        if os.path.isdir(f): constants.safe_delete(f)
+                        else:                os.remove(f)
 
                     # Log issues to console during debug
                     except Exception as e:
-                        if constants.debug:
-                            print(f'Error deleting file/folder during restore: {e}')
+                        send_log('restore_server', f'Error deleting file/folder during restore: {e}', 'error')
 
 
 
@@ -477,7 +487,11 @@ def restore_server(name: str, backup_name: str, backup_stats=None):
                 constants.server_manager.check_for_updates()
 
                 send_log('restore_server', f"successfully restored '{name}' to '{backup_name}'", 'info')
-                return [os.path.basename(backup_name), convert_size(os.stat(file_path).st_size), convert_date(os.stat(file_path).st_mtime)]
+                return {
+                    'name': os.path.basename(backup_name),
+                    'size': convert_size(os.stat(file_path).st_size),
+                    'date': convert_date(os.stat(file_path).st_mtime)
+                }
 
         except Exception as e:
             send_log('restore_server', f"error restoring '{name}' to '{backup_name}': {constants.format_traceback(e)}", 'error')
@@ -488,9 +502,15 @@ def set_backup_directory(name: str, new_dir: str, new_amount: str):
 
     cwd = constants.get_cwd()
     config_file = manager.server_config(name)
-    current_dir = config_file.get('bkup', 'bkupDir')
-    current_dir = current_dir.replace(r"/","\\") if constants.os_name == 'windows' else current_dir
-    new_dir = new_dir.replace(r"/","\\") if constants.os_name == 'windows' else new_dir
+    try:
+        current_dir = config_file.get('bkup', 'bkupDir')
+        current_dir = current_dir.replace(r"/","\\") if constants.os_name == 'windows' else current_dir
+        new_dir = new_dir.replace(r"/","\\") if constants.os_name == 'windows' else new_dir
+
+    except Exception as e:
+        send_log('set_backup_directory', f'error migrating back-up directory:\n{constants.format_traceback(e)}', 'error')
+        return None
+
 
     if set_lock(name, True, 'migrate'):
 
@@ -517,10 +537,8 @@ def set_backup_directory(name: str, new_dir: str, new_amount: str):
                         configs = glob(os.path.join(extract_folder, 'auto-mcs.ini'))
                         configs.extend(glob(os.path.join(extract_folder, '.auto-mcs.ini')))
                         for cfg in configs:
-                            config = ConfigParser(allow_no_value=True, comment_prefixes=';', interpolation=None)
-                            config.optionxform = str
-                            config.read(cfg)
-                            if config:
+                            config = load_config(cfg)
+                            if config.sections():
                                 if config.get('general', 'serverName') == name:
 
                                     # Update bkupDir with new_dir in the back-ups' auto-mcs.ini
@@ -559,11 +577,17 @@ def set_backup_directory(name: str, new_dir: str, new_amount: str):
 # Migrate backup names when server is renamed
 def rename_backups(name: str, new_name: str):
     if set_lock(name, True, 'migrate'):
-        config_file = manager.server_config(new_name)
-        current_dir = config_file.get('bkup', 'bkupDir')
-        current_dir = current_dir.replace(r"/", "\\") if constants.os_name == 'windows' else current_dir
-        file_list   = glob(os.path.join(current_dir, f"{name}__*"))
-        failure     = False
+        try:
+            config_file = manager.server_config(new_name)
+            current_dir = config_file.get('bkup', 'bkupDir')
+            current_dir = current_dir.replace(r"/", "\\") if constants.os_name == 'windows' else current_dir
+            file_list   = glob(os.path.join(current_dir, f"{name}__*"))
+            failure     = False
+
+        except Exception as e:
+            send_log('rename_backups', f'error renaming all back-ups for:\n{constants.format_traceback(e)}', 'error')
+            return None
+
 
         if file_list:
             send_log('rename_backups', f"renaming all back-ups for '{name}' to '{new_name}'...", 'info')
@@ -605,11 +629,8 @@ def rename_backup(file: str, new_name: str):
         config_files.extend(glob(os.path.join(extract_folder, '.auto-mcs.ini')))
 
         for cfg in config_files:
-            config = ConfigParser(allow_no_value=True, comment_prefixes=';', interpolation=None)
-            config.optionxform = str
-            config.read(cfg)
-
-            if config:
+            config = load_config(cfg)
+            if config.sections():
                 if config.get('general', 'serverName') == name:
 
                     # Update bkupDir with new_dir in the back-ups' auto-mcs.ini
